+++ CHANGELOG.rst
@@ -1,5 +1,11 @@
 Changelog
 =========
+* 0.9.5 (November 1, 2023)
+    * Added the ability to void all voidable QB types
+    * Added to_ref to CreditMemo object
+    * Added ProjectRef and ShipFromAddr to Estimate
+    * Added missing initialization for objects on DiscountLineDetail, Estimate, Employee, and Invoice
+
 * 0.9.4 (August 29, 2023)
     * Removed python 2 compatible decorators
     * Removed python 2 dependencies